# scooter/swagger.yaml
swagger: '2.0'
info:
  title: SCOOTER API REST
  description: It is API to scooter app
  termsOfService: https://www.google.com/policies/terms/
  contact:
    email: misael.gonzalez.e.229@gmail.com
  license:
    name: BSD License
  version: v1
basePath: /
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Basic:
    type: basic
  bearerAuth:
    type: http
    scheme: bearer
    bearerFormat: JWT
security:
  - Basic: []
  - bearerAuth: []

# ====== PATHS =======
paths:
...
  /api/v1/users/forgot-password/:
    post:
      operationId: forgot_password_customer
      description: Correo para obtener token de recuperación
      parameters:
        - name: body
          in: body
          description: Recuperar contraseña
          schema:
            $ref: '#/definitions/ForgotPassword'
      responses:
        '200':
          description: 'Envío de correo para obtener un token de recuperación'
      tags:
        - Clientes
    parameters: []
  # RECOVER PASSWORD
  /api/v1/users/recover-password/:
    post:
      operationId: recover_password_customer
      description: Recuperar contraseña del cliente
      parameters:
        - name: body
          in: body
          description: Recuperar contraseña
          schema:
            $ref: '#/definitions/RecoverPassword'
      responses:
        '200':
          description: 'Contraseña cambiada'
      tags:
        - Clientes
    parameters: []
  # REGISTER CUSTOMER
...
definitions:
...
  ForgotPassword:
    type: object
    properties:
      email:
        type: string
        description: Correo
    required:
      - email
  # RECOVER PASSWORD
  RecoverPassword:
    type: object
    properties:
      token:
        type: string
        description: Token de recuperación
      password:
        type: string
        description: Nueva contraseña
    required:
      - token
      - password
  # REGISTER CUSTOMER
...